v2.04 - fixed music re-play problem

+++ src/DirectDraw.cpp
@@ -960,7 +960,7 @@ VOID DirectDraw::RenderOld()
 	MemoryFree(pixelBuffer);
 }
 
-VOID __fastcall UseShaderProgram(ShaderProgram* program, FLOAT texSize)
+VOID __fastcall UseShaderProgram(ShaderProgram* program, DWORD texSize)
 {
 	if (!program->id)
 	{
@@ -991,7 +991,7 @@ VOID __fastcall UseShaderProgram(ShaderProgram* program, FLOAT texSize)
 
 		loc = GLGetUniformLocation(program->id, "texSize");
 		if (loc >= 0)
-			GLUniform2f(loc, texSize, texSize);
+			GLUniform2f(loc, (FLOAT)texSize, (FLOAT)texSize);
 	}
 	else
 		GLUseProgram(program->id);
